remove task output configuration
Now all tasks store output according to their name. This means that
once a value is set, it will not be changed and can safely be referred
to later. This also means that there can be no input with the same name
as a task, which I will sort out later...